LnProb function and rho move edits

+++ src/likelihood.h
@@ -145,9 +145,9 @@ int       Likelihood_Res_SSE (TreeNode *p, int division, int chain, MrBFlt *lnL,
 #endif
 int       Likelihood_Std (TreeNode *p, int division, int chain, MrBFlt *lnL, int whichSitePats);
 int       Likelihood_StdCorr (TreeNode *p, int division, int chain, MrBFlt *lnL, int whichSitePats);
-int       LnProbAllocation (int *allocationVector, int numChars, MrBFlt alphaDir);
-int       LnProbLatentCluster (int *latentColumn, int allocationValue, int numChars, int *allocationVector, int chain);
-int       LnProbLatentMatrix (int *allocationVector, int *latentMatrix, int numChar, int chain);
+double    LnProbAllocation (int *allocationVector, int numChars, MrBFlt alphaDir);
+double    LnProbLatentCluster (int *latentColumn, int allocationValue, int numChars, int *allocationVector, int chain);
+double    LnProbLatentMatrix (int *allocationVector, int *latentMatrix, int numChar, int chain);
 int       TiProbs_Fels (TreeNode *p, int division, int chain);
 int       TiProbs_Gen (TreeNode *p, int division, int chain);
 int       TiProbs_GenCov (TreeNode *p, int division, int chain);
